# swagger.yml
...
swagger: "2.0"
info:
  title: MinIO Console Server
  version: 0.1.0
consumes:
  - application/json
produces:
  - application/json
schemes:
  - http
  - ws
basePath: /api/v1
...
securityDefinitions:
  key:
    type: oauth2
    flow: accessCode
    authorizationUrl: http://min.io
    tokenUrl: http://min.io
  anonymous:
    name: X-Anonymous
    in: header
    type: apiKey
# Apply the key security definition to all APIs
security:
  - key: [ ]
...
paths:
  /login:
    get:
      summary: Returns login strategy, form or sso.
      operationId: LoginDetail
      responses:
        200:
          description: A successful response.
          schema:
            $ref: "#/definitions/loginDetails"
        default:
          description: Generic error response.
          schema:
            $ref: "#/definitions/ApiError"
      # Exclude this API from the authentication requirement
      security: [ ]
      tags:
        - Auth
    post:
      summary: Login to Console
      operationId: Login
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: "#/definitions/loginRequest"
      responses:
        204:
          description: A successful login.
        default:
          description: Generic error response.
          schema:
            $ref: "#/definitions/ApiError"
      # Exclude this API from the authentication requirement
      security: [ ]
      tags:
        - Auth
  /logout:
    post:
      summary: Logout from Console.
      operationId: Logout
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: "#/definitions/logoutRequest"
      responses:
        200:
          description: A successful response.
        default:
          description: Generic error response.
          schema:
            $ref: "#/definitions/ApiError"
      tags:
        - Auth
...
definitions:
...
  ApiError:
    type: object
    properties:
      message:
        type: string
      detailedMessage:
        type: string
...
  loginDetails:
    type: object
    properties:
      loginStrategy:
        type: string
        enum: [ form, service-account, redirect-service-account ]
      redirectRules:
        type: array
        items:
          $ref: "#/definitions/redirectRule"
      isK8S:
        type: boolean
      animatedLogin:
        type: boolean
  loginRequest:
    type: object
    properties:
      accessKey:
        type: string
      secretKey:
        type: string
      sts:
        type: string
      features:
        type: object
        properties:
          hide_menu:
            type: boolean
  logoutRequest:
    type: object
    properties:
      state:
        type: string
  # Structure that holds the `Bearer {TOKEN}` present on authenticated requests
...
  redirectRule:
    type: object
    properties:
      redirect:
        type: string
      displayName:
        type: string
      serviceType:
        type: string
...